feat: nix-based floxenv builder
This commit adds the flox-buildenv package to the build, which is
an adaptation of the upstream nixpkgs buildenv builder altered to
produce multiple outputs:

1. "out": the default, created exactly like the nixpkgs buildenv
2. "develop": the flox develop environment, populated with all
   propagated-build-inputs treated as if they were propagated-
   user-env-packages, and thus recursively linked
3+ "build-<pname>": a flox environment comprised solely from the
   environment's "toplevel" package group, and optionally further
   filtered based on the "build.<pname>.packages" manifest attribute.

Each of the above outputs will also automatically include the flox
activation scripts ("etc-profile") and manifest personalization packages.

The bulk of the business logic resides in the `buildenv.nix` and `builder.pl`
files, but the buildenv package also includes two transition scripts that have
been helpful in development:

- `bin/buildenv`: wrapper script for invoking `nix build` against `buildenv.nix`
- `bin/pkgdb`: arg-compatible `pkgdb` replacement which invokes `bin/buildenv`

Both of the above scripts can be _deleted_ once the rust code has been
modified to call `nix build ... buildenv.nix` directly.

Finally, you'll note that this commit includes both a copy of `builder.pl`
modified to parse the flox manifest, and a patch for rendering that modified
version from the nixpkgs upstream version. The work to parse the manifest and
calculate the stream of packages to be linked is currently being done
_imperatively_ in the perl script, but it could similarly be done
_functionally_ from within `buildenv.nix`.  If we were to elect to do the
latter then the patch would become much shorter, and may be preferable if we
are concerned about maintaining perl code.