refactor: error for nested activate in Rust
We currently error for an interactive `flox activate` in Bash. With
upcoming activation changes, Bash won't be looking at
`_FLOX_ACTIVE_ENVIRONMENTS`, it will only decide whether to start or
attach based on whether an environment is already activated somewhere,
which we shouldn't error for.

Move the error to Rust, which is also consistent with the activation
design diagram.

Also add a test as this behavior was not previously covered.