fix: use shorter activation id and create dir
We currently use Uuid as the id of a flox activation. Since we may be
using this in the path for sockets, we should keep it shorter than a
uuid. We can instead use the first 8 characters of the hash of the nix
store path for an activation.

We should also create the activation state directory in flox-activations
so we don't have to create it in the activation script.